# openapi/openapi.yaml
openapi: 3.0.0
info:
  title: nijisongs-api
  description: にじさんじライバーの歌動画とライバー情報を扱う
  version: 1.0.0
servers:
  - url: http://localhost:8081
    description: Development server
paths:
  /youtube/search:
    get:
      tags:
        - youtube
      summary: 動画ID取得
      description: Youtube Data API を使って、にじさんじの歌動画に関連する動画IDを取得する<br> publishedは動画がアップロードされた時間(公開日ではない)<br>
        publishedの形式は(2021-01-01T00:00:00Z)<br>
        publishedを指定しない場合、実行したその日の(JST)0:00~23:59の時間帯にアップロードされた動画IDを取得する
      parameters:
        - name: publishedAfter
          in: query
          description: 指定された日時より後にアップロードされた動画を取得
          schema:
            type: string
        - name: publishedBefore
          in: query
          description: 指定された日時より前にアップロードされた動画を取得
          schema:
            type: string
      responses:
        '200':
          description:
            publishedAfter=2021-01-01T00:00:00Z<br>
            publishedBefore=2021-01-02T06:00:00Z の場合
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
                example:
                  - 8lWZpNjntU8
                  - yBU8FDYR3_Q
        '500':
          description:
            Youtube Data API の取得に失敗した時
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
          
  /youtube/activities:
    get:
      tags:
        - youtube
      summary: 動画ID取得
      description: 
        指定されたチャンネルと期間から投稿された動画IDを取得する<br>
        datetimeを両方指定しない場合、現在から7時間前の期間を取得する<br>
        原因が分からないが、プレミアム公開(公開予約)の動画IDの取得ができないため、既に公開された動画IDを取得する際に使用した方が良い
      parameters: 
        - name: channelId
          in: query
          required: true
          description: コンマ区切りで指定可能(UCX7YkU9nEeaoZbkVLVajcMg, ...)
          schema:
            type: string
        - name: datetimeAfter
          in: query
          description: データ形式(2021-01-01T00:00:00Z)
          schema:
            type: string
        - name: datetimeBefore
          in: query
          description: データ形式(2021-01-01T00:00:00Z)
          schema:
            type: string
      responses:
        '200':
          description: channelId=UCX7YkU9nEeaoZbkVLVajcMg datetimeAfter=2021-01-01T00:00:00Z datetimeBefore=2021-01-02T00:00:00Z の場合
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
                example:
                  - IGvDbjMukdQ
                  - 0fLDi5tlYew
        '500':
          description:
            Youtube Data API の取得に失敗した時
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  
  /youtube/videos:
    get:
      tags:
        - youtube
      summary: 動画詳細データを取得
      description:
        動画IDから動画の詳細データを取得する<br>
        また、歌動画か判定されたデータ(songConfirm)と歌動画かもしれないデータ(unSongConfirm)だけ取得することも可能<br>
      parameters: 
        - name: videoId
          in: query
          required: true
          schema:
            type: string
        - name: select
          in: query
          description: 歌動画か判定するのか(select指定するにはpartを指定しない or partにsnippetを指定する必要がある)
          schema:
            type: boolean
        - name: part
          in: query
          description: 取得するデータの部分を指定(snippet, contentDetails, statistics, ...) 指定しない場合、全てのデータを取得する
          schema:
            type: string
      responses:
        '200':
          description: 
            select=trueにするとsongConfirm[], unSongConfirm[]が取得できる<br>
            select=false(or 指定しない)とresultのみ取得できる
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  result:
                    type: array
                    items:
                      $ref: '#/components/schemas/youtube'
                  songConfirm:
                    type: array
                    items:
                      $ref: '#/components/schemas/youtube'
                  unSongConfirm:
                    type: array
                    items:
                      $ref: '#/components/schemas/youtube'
        '500':
          description:
            Youtube Data API の取得に失敗した時
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                      
  /videos:
    get:
      tags:
        - video
      summary: 動画詳細データを取得
      description:
        DBに保存されている動画詳細データを取得する<br>
        データの順番は再生回数が多い順に取得される
      parameters:
        - name: id
          in: query
          description: 取得したい動画IDを指定　コンマ区切りで複数指定できる
          schema:
            type: string
        - name: songConfirm
          in: query
          description: 歌動画(true)か歌動画かもしれない(false)データのみ取得することができる　指定しない場合、全てのデータを返す
          schema:
            type: boolean
        - name: maxResult
          in: query
          description: 取得するデータの数を指定できる　指定しない場合、全てのデータを返す
          schema:
            type: integer
        - name: page
          in: query
          description: 取得するデータのページを指定できる　指定しない場合、最初のページのデータを返す
          schema:
            type: integer
        - name: startAtAfter
          in: query
          description:
            指定された時間より後に公開するデータを取得<br>
            データ形式(2021-01-01T00:00:00Z)
          schema:
            type: string
        - name: startAtBefore
          in: query
          description:
            指定された時間より前に公開するデータを取得<br>
            データ形式(2021-01-01T00:00:00Z)
          schema:
            type: string
      responses:
        '200':
          description: 取得結果
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/video'
        '500':
          description:
            取得に失敗した時
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
    post:
      tags:
        - video
      summary: 動画詳細データを保存
      description:
        動画詳細データをDBに保存する<br>
      requestBody: # リクエストボディ
        description: youtube/videosから取得したデータ
        content:
          application/json:
            schema:
              type: object
              properties:
                message:
                  type: string
                result:
                  type: array
                  items:
                    $ref: '#/components/schemas/youtube'
                songConfirm:
                  type: array
                  items:
                      $ref: '#/components/schemas/youtube'
                unSongConfirm:
                  type: array
                  items:
                    $ref: '#/components/schemas/youtube'
      responses:
        '201':
          description: success
        '500':
          description:
            保存に失敗した時
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
    
    put:
      tags:
        - video
      summary: 動画詳細データの更新
      description:
        動画詳細データが、歌ってみた動画であることが確定した時、songConfirmをTrueに更新できる<br>
        また、動画のタイトル、概要欄を更新できる
      requestBody:
        description: videoIdは動画ID
        content:
          application/json:
            schema:
              type: object
              properties:
                videoId:
                  type: string
                songConfirm:
                  type: boolean
                title:
                  type: string
                description:
                  type: string
      responses:
        '201':
          description: 更新に成功すると、動画詳細データの一部を返す
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  title:
                    type: string
                  description:
                    type: string
                  startTime:
                    type: string
                  createdAt:
                    type: string
                  songConfirm:
                    type: boolean
        '500':
          description:
            更新に失敗した時
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
    
    delete:
      tags:
        - video
      summary: 動画詳細データの削除
      parameters:
        - name: id
          in: query
          required: true
          description: 削除したい動画のIDを指定　コンマ区切りで複数指定できる
          schema:
            type: string
      responses:
        '201':
          description: DELETED
        '500':
          description:
            削除に失敗した時
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
  
  /videos/viewCount:
    put:
      tags:
        - video
      summary: 動画の視聴回数を更新
      description:
        歌動画の視聴回数をYoutube APIから取得し、DBに保存されている動画詳細データの視聴回数を上書きする
      responses:
        '201':
          description: updated
          
  /vtuber:
    get:
      tags:
        - vtuber
      summary: Vtuber情報を取得
      description:
        DBに保存してあるVtuberデータを取得<br>
        パラメータを指定することで、その条件に合ったデータのみ取得することができる<br>
        指定しない場合、全てのデータを返す
      parameters:
        - name: affiliation
          in: query
          description: 所属を指定できる(にじさんじ, にじさんじKR, にじさんじ卒業, ...)
          schema:
            type: string
        - name: name
          in: query
          description: 名前を指定できる(完全一致な名前)
          schema:
            type: string
        - name: id
          in: query
          description: チャンネルIDを指定
          schema:
            type: string
      responses:
        '200':
          description: 取得結果
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/vtuber'
        '500':
          description:
            取得に失敗した時
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  
    post:
      tags:
        - vtuber
      summary: 動画詳細データを保存
      description:
        動画詳細データをDBに保存する<br>
      requestBody:
        description: user to create
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: string
                name:
                  type: string
                readname:
                  type: string
                affiliation:
                  type: string
                birthday:
                  type: string
                image:
                  type: string
      responses:
        '201':
          description: success
        '500':
          description:
            保存に失敗した時
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
    
    put:
      tags:
        - vtuber
      summary: vtuber情報を編集
      description: 編集したい部分のみ指定できる
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                channelId:
                  type: string
                name:
                  type: string
                readname:
                  type: string
                affiliation:
                  type: string
                birthday:
                  type: string
                image:
                  type: string
                  
      responses:
        '201':
          description: UPDATED
        '500':
          description:
            更新に失敗した時
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
          
  /tags:
    get:
      tags: 
        - tags
      summary: tag情報を取得
      description:
        DBに保存してあるTagデータを取得<br>
      parameters:
        - name: names
          in: query
          description: 取得したいtag名を指定(コンマ区切りで複数指定できる)
          schema:
            type: string
      responses:
        '200':
          description: 取得結果
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GETtags'
        '500':
          description:
            取得に失敗した時
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
  
    post:
      tags:
        - tags
      summary: tag情報を保存
      description:
        videoIdに関連するtag情報を保存する<br>
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                video_tags:
                  type: array
                  items:
                    type: object
                    properties:
                      videoId:
                        type: string
                      tags:
                        type: object
                        properties:
                          name: 
                            type: string
                          description:
                            type: string
      responses:
        '201':
          description: success
        '500':
          description:
            保存に失敗した時
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
      
    delete:
      tags: 
        - tags
      summary: tag情報を削除
      description:
        DBに保存してあるTagデータを削除<br>videoIdのみ指定した場合、指定された動画の全てのタグが削除される
      parameters:
        - name: names
          in: query
          description: 削除したいtag名を指定(コンマ区切りで複数指定できる)
          schema:
            type: string
        - name: videoId
          in: query
          description: videoIdから削除したいtag名を指定できる
          
          schema:
            type: string
      responses:
        '201':
          description: success
        '500':
          description:
            削除に失敗した時
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
          
components:
  schemas:
    youtube:
      type: object
      properties:
        kind:
          type: string
        etag:
          type: string
        id:
          type: string
        snippet:
          type: object
        contentDetails:
          type: object
        statistics:
          type: object
        liveStreamingDetails:
          type: object
            
    statistic:
      type: object
      properties:
        id:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
        viewCount:
          type: integer
        likeCount:
          type: integer
        dislikeCount:
          type: integer
        commentCount:
          type: integer
    
    thumbnail:
      type: object
      properties:
        id:
          type: string
        defaultUrl:
          type: string
        medium:
          type: string
        high:
          type: string
        standard:
          type: string
        maxres:
          type: string
          
    tags:
      type: object
      properties:
        description:
          type: string
        tag:
          type: object
          properties:
            name:
              type: string
        
    video:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        startTime:
          type: string
        createdAt:
          type: string
        songConfirm:
          type: boolean
        thumbnail:
          $ref: '#/components/schemas/thumbnail'
        statistic:
          $ref: '#/components/schemas/statistic'
        tags:
          type: array
          items:
            $ref: '#/components/schemas/tags'
    
    vtuber:
      type: object
      properties:
        id: 
          type: string
        name:
          type: string
        readname:
          type: string
        affiliation:
          type: string
        birthday:
          type: string
        image:
          type: string
    
    GETtags:
      type: object
      properties:
        id: 
          type: integer
        name:
          type: string
        publish:
          type: boolean
        
tags:
  - name: youtube
    description: Access to youtube
  - name: video
    description: Access to video
  - name: vtuber
    description: Access to vtuber
  - name: tags
    description: Access to tags